fix: resolve PR commets
