+++ src/pages/MainPage/ui/CardSkeleton.tsx
